ast: Implement eqfilter and path
The communcation between the parser and the resolver is going to be done
using a Abstract Syntax Tree or AST for short, also the idea is to start
representing an expression a filter expression like
`routes.running.destination=="0.0.0.0/0"`.

This change add the needed types to create a tree to represent that with
the less priority operator at the top, it contains some unit test to
check the yaml annotations.

Signed-off-by: Quique Llorente <ellorent@redhat.com>