expression, error: Decorate errors with a snippet
To debug errors is good to have the capture entry expression that is
generating then and also the point within the expression. This change
add a "expression" package with a error wrapper that add a position and
is able to "decorate" the error with a snippet that mix the expression
and the position.

Signed-off-by: Quique Llorente <ellorent@redhat.com>