resolver, name: Add resolve by name
For resolve stuff like "capture.default-gw" the capture expression will
have to be resolved by it's name and also evaluated in place in case
is not already being evaluated. This change add a function to find the
resolved expressions in the map of results and if it's not there resolve
in place and store it at the results map.

Signed-off-by: Quique Llorente <ellorent@redhat.com>