resolver: Pass capturedState as argument
On some scenarios some capture entry name not present at the capture are
referenced and they have to return the content at cached captured state.
This change pass the capturedState from the cache to the resolver.

Also as an effort to keep the original design, this change convert the public
structs for lexer, parser and resolver as stateless.

Signed-off-by: Quique Llorente <ellorent@redhat.com>