capture: Return unmarshaled capturedStates too
The function ResolveCaptureEntryPath need the capturedState to resolve
the capture entry, if this capturedState need to be the unmarshaled one
or a new extra step is needed to unmarshal capturedStates. This change
creates resolver.Result to return the marshaled and unmarshaled
capturedStates and also move CaptureEntry to Result struct so there is a
capture.Result too that will contain the ResolveCaptureEntryPath and
will access the unmarshaled desiredStates from the resolverResult field.

Signed-off-by: Quique Llorente <ellorent@redhat.com>