internal, types: Create the internal types
This change create the types that will be used internally by nmpolicy,
they represent the unmarshaled version of the nmstate wrapped with a
struct similar to the exported types.

Also some types declarations has being added to make code more readable
so the monstruosity "map[string]map[string]interface{}" and the like is
no longer used.

Signed-off-by: Quique Llorente <ellorent@redhat.com>