resolver: Implement replace operation
As first step to implement the Replace operation at the resolver this
change implement path value resplacement but just for current state
since it does not involve changing path.go file to traverse capture
entry path ref and apply a function.

The replace operation has to be able to use a capture ref path as input,
also some tuning is needed at the apply functions since the modified
value has to be returned and that's different from filter function.

Signed-off-by: Quique Llorente <ellorent@redhat.com>