resolver: Change return at resolverFilteredValue
The resolver helper functions should resolve ast.Node and return a
interface{} or specific type, the resolverFilteredValue was returning an
ast.Node and also generating with the value of the capture entry path.
This change use interface{} as the return value from
resolverFilteredValue and also doing so remove the limitation of types
we can filter/replace.

Signed-off-by: Quique Llorente <ellorent@redhat.com>