resolver: Replace apply with visitor pattern
The resolver apply functions use a little of all the arguments that they
pass between each other, most of those arguments can be some state at
some visitor struct. This change replace the apply functions with a
visitor pattern consisting of struct with proper state and a visit
function per type it visits.

Signed-off-by: Quique Llorente <ellorent@redhat.com>