lexer, parser: Decorate errors with snippet
At lexer, parser is easy to detect the rune, token that is generating
the error. This change wrap those errors with the expression and a
pointer.

Since the parser knows only about tokens, this change add also the
expression to the parser so it can use it to wrap the errors.

Signed-off-by: Quique Llorente <ellorent@redhat.com>