resolver: Wrap errors with expressions
The ast.Node contains a Position field that points to the expressions
rune where the ast.Node was generated from. This change use that
position to to wrap the errors with a snippet that contains the
expression and a pointer to the place that generated the error.

To acomplish that the capture expressions are passed to the resolver so
it can search for the expression string to create the snippet.

The snippet pointer marks until operators but not inside the paths, this
can be fixed at a follow up.

Signed-off-by: Quique Llorente <ellorent@redhat.com>