types: Add custom marhsaling for desiredState
Representing yaml inside yaml is kind of problematic so the yaml parser
has to be instructed just ignore those fields and store them as []byte.
This change uses MarshalJSON/UnmarshalJSON functions to do that, also
they will need a specific NMState struct so we can put those receivers
on it.

Signed-off-by: Quique Llorente <ellorent@redhat.com>