ci: Github actions and netifly preview for docs
For the documentation we need a pair of things from CI, preview the PR
as a webpage to check how it looks and publish it on a github branch so
it can be viewed as a github pages domain.

This change add some netifly.yaml to preview the documentation PR and
some github actions to build publish the page.

Signed-off-by: Quique Llorente <ellorent@redhat.com>