resolver: Express should filter with a step map
The filter resolver do filtering from `interfaces` or `routes.running`
always using a pair of flags that state that are changing when the first
time a slice is encountered, this behaviour is not very clear. This
change make it explicit using a map with steps it has to filter, it also
alow us to remove the pathVisitor copy at visitSlice.

Signed-off-by: Quique Llorente <ellorent@redhat.com>