replace: Fix issue with setting value to a sub-key, that doesn't exist
This commit constructs map key according to a path given to replace
operator, so that replace may set a value to a sub-key, that doesn't
exist in the original state.

For an example, consider the following state:
state:
  interfaces:
  - name: eth0
    type: ethernet

If we attempt to use replace to set interfaces.lldp.enabled:=true, the
following result is expected with the chanes in this commit:
state:
  interfaces:
  - name: eth0
    type: ethernet
    lldp:
      enabled: true

Without constructing the map structure, replace returned nil.
Only maps are constructed in thie fashion, replace in a missing slice
remains unsupported.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>